[rename]: #22 shared/ui에 modal 폴더로 파일 정리
