+++ src/shared/ui/index.ts
@@ -1,3 +1,3 @@
 export * from "./Header";
-export * from "./QrModal";
-export * from "./QrSection";
+export * from "./modal/QrModal";
+export * from "./modal/QrSection";
