+++ src/shared/ui/modal/QrModal.tsx
