+++ src/shared/ui/modal/QrSection.tsx
